rename: 컴포넌트 파일 이름 PascalCase로 변경
