+++ src/app/layout.tsx
@@ -1,4 +1,4 @@
-import Container from "@/components/common/container";
+import Container from "@/components/common/Container";
 import Header from "@/components/layout/header";
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
