+++ src/components/layout/header/SearchIcon.tsx
